Multiple Changes to Kinks Page
Had to get the page into a format consistent with the rest of the site. I know this wasn't Sarah's original intention but there was no CSS file included in the pull request, so I did what I could. Will meet with Sarah so we can get her vision on the page.